Update CVE parsing logic to handle "rejected" CVEs
When a CVE is marked as "Rejected" on the CVE page, the HTML structure
is quite different than on a normal page. This updates the logic to be
able to parse that style of page, and check for the Rejected HTML
structure if the normal structure isn't present. If neither is present,
the resource will now output to the log and panic. This should be a
better guard against future changes to the HTML structure.

[#179154907] cloudfoundry/usn-resource #2: [Snyk] Security upgrade alpine from 3.4 to 3

Signed-off-by: Brian Upton <bupton@vmware.com>