Handle CVE pages missing priority
Some CVE pages seem to be missing the priority badge that is normally on
pages. To allow this case, we carefully check that page structure
matches what we expect and return an "unknown" status. If it doesn't
match, we fall through to the previous code that causes a failure. An
existing integration test will fail if all CVE priorities return as
unknown, which should also protect against our parser breaking.

Example page: https://ubuntu.com/security/CVE-2021-25219

Additionally adds a "not in ubuntu" status that is now a potential
priority

[#180169622] Fix USN Resource