USN resource now understands brotli 🥦
One consequence of this change is that the code in http_client.go now
returns `io.Reader, int, error` instead of `http.Response, error` which
is a consequence of how the bortli decoder expects to be called.
Internally the http_client.go code reads the response.Body and, if
necessary, passes it through the brotli decoder, and finally creates a
reader with the (formerly brotli encoded or not) bytes to pass back to
the caller.

See: https://medium.com/sellerapp/decoding-brotli-in-golang-e024dcf34be1

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>