Corrects selector for pages for Rejected CVEs
No clue if this was a typo or if the element's class was really spelt
as "p-heading--four", but it definitely is not anymore.

<https://ubuntu.com/security/CVE-2021-20223> is a CVE that has recently
been Rejected, and cause the tests for 'usn-resource' to fail. We have
verified that CVE-2021-20223's page structure matches the change we're
making here, so hopefully this will work forever and always.

Unfortunately, the tests for usn-resource use live site data, so there's
no guarantee that we'll always process a page with a `Rejected` CVE.

Signed-off-by: Chris Selzo <chris.selzo@broadcom.com>