Include previous resource version in check
Concourse expects you to include the previous latest version of the
resource in the array returned by `check` "if it's still valid."
Previously, we were only including new versions.

It does not seem to be explicitly documented what Concourse will do if
you do not include the previous latest version in the array, but we've
seen evidence of versions going missing and reports of USNs / CVEs not
appearing in the release notes.

Signed-off-by: Aram Price <aram.price@broadcom.com>